Add auto archive feature

+++ db/schema.rb
@@ -27,7 +27,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_11_17_215855) do
     t.string "name", null: false
     t.text "body"
     t.string "record_type", null: false
-    t.bigint "record_id", null: false
+    t.integer "record_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["record_type", "record_id", "name"], name: "index_action_text_rich_texts_uniqueness", unique: true
